This is the initial import of prototype USB drivers written for the OHCI chip on the customer L board.
Details:
  The drivers are for mouse, keyboard and combined keyboard and mouse.  They
assume that the correct device is going to be plugged in, not reading the
descriptors at all.  Once the mouse is plugged in, it will become the active
pointer device, reverting to the previous one when extracted.

Admin:
  The prototypes have been tested on a SDB customer L board, but require a couple
of pokes to the reset and configuration register to get them going:
  memorya f99f611c 3840 works on the SDB but will be different on the customer L board