Now handles interrupt out transfers
Detail:
  There were many places where this was broken: the devicefs
interface was always starting off a read regardless of
whether it was an IN or OUT transfer, usbdi.c was setting
OUT interrupt transfers as repeating, which doesn't make sense,
and ohci.c was always making interrupt transfers OUT (this is
fixed in the latest BSD sources). Also the code which
unhalts endpoints at start of day has been removed, as this is
no longer in the BSD sources having been removed after it was
seend to have caused problems.

Admin:
  Tested with a CATC analyzer on an AverMedia FM radio. If now
operates properly with the code from Dave Higton.

Version 0.43. Tagged as 'NetBSD-0_43'