Internationalisation, help, and syntax improvments
Assign a proper error number and translate "Incompatible USBDriver version".
Don't bother internationalising the *commands which are for debug builds only.
Rephrase the help for USBQuirk so it makes sense, and trim the unmatched '<' from the syntax example.
Bracket the USBDiscover help with the same switch that governs the code, which has been disabled for over a decade.

Version 1.04. Tagged as 'NetBSD-1_04'