Align docs with alias variable reality
The documentation of the Alias$@USBDevice variables barely made sense, and examples not that consistent - improve this.
Fix step 1 of the alias handling to match the docs (this was fixed previously in revision 1.34 but accidentally overwritten in 1.35 less than an hour later).
Add a check for malloc() failing in low memory situations.

Tested with a couple of USB devices and alias variables containing 'echo' commands.

Version 1.22. Tagged as 'USBDriver-1_22'