Adapt to different `C_EXP_HDR` settings
An upcoming change to BuildSys adjusts the default setting of `C_EXP_HDR` so
that it doesn't include a trailing `.h` for native builds. Since we use the
value of this variable in one of our rules, adapt accordingly (note that this
implementation will work with both old and new BuildSys).

Version 1.31. Not tagged